Added code to split up monthly data by router
Added code to export_data.py which will split up the sqlite3 version
of the monthly database into a bunch of smaller databases separated
by the router. This is necessary 1) to do multitheading analysis
(sqlite does not support concurrency) and 2) to improve performance
by reducing the db size. This is a first step to running these
computations in parallel